Remove explicit disable for the Piro package.
But keep the explicit disables (for now) for the two test cases that were failing.
The issue was documented here https://github.com/trilinos/Trilinos/issues/11169,
and its solution is in PR https://github.com/trilinos/Trilinos/pull/11185. I am
pushing just the remove of the disable for now for our push to the C++17 standard
so we're not blocked and at the same time we don't lose coverage completely for
the Piro package and its dependencies.